feature/sc-1195/forbid-user-enter-login-signup-page-after (#136)
* Refactor authentication redirect as custom hook

Created a custom hook useRedirectIfAuthenticated to handle redirecting users if they are already authenticated. This hook uses useSession from next-auth/react and useRouter from next/router to check the user's authentication status and redirect them to the home page if they are already authenticated.

* spacing

* feat: add loading indicator to layout logic

Add new logic to determine whether to show a loading indicator in the layout based on the `isLoading` state. If `isLoading` is true, the loading indicator will be displayed. Otherwise, the content of the layout will be displayed.

Fixes #123

* feat: add redirect logic to AuthLayout component

Add new logic to determine whether to redirect the user to the login page if they are not authenticated. If the user is not authenticated and the `shouldRedirect` prop is true, they will be redirected to the login page. Otherwise, the component will render its children as usual.

* deleted the rauth hook